+++ application/models/Robotsdata.php
@@ -11,37 +11,55 @@
  *
  * @author Jake
  */
-class Robotsdata extends CI_Model {
+class Robotsdata extends CI_Model
+{
 
     // Constructor    
-    public function __construct() {
+    public function __construct()
+    {
         parent::__construct();
     }
 
     // retrieve all of bots
-    public function getAllBots() {
+    public function getAllBots()
+    {
         return $this->bots;
     }
+	public function getAllBotsAsArray(){
+		return $this->db->get('assembledbots')->result_array();
+	}
 
     // retrieve a single bot
-    public function getBot($which) {
+    public function getBot($which)
+    {
         // iterate over the data until we find the one we want
-        foreach ($this->bots as $record) {
+        foreach ($this->db->get('assembledbots')->result_array() as $record) {
             if ($record['id'] == $which) {
+
                 return $record;
             }
         }
-          return null;
+        return null;
     }
     
-    public function createBot($data) {
+	function size()
+	{
+		$query = $this->db->get('assembledbots');
+		return $query->num_rows();
+	}
+	
+   
+
+    public function createBot($data)
+    {
         $this->db->insert('assembledbots', $data);
     }
-    
+
     // Returns an ID number by adding 1 to the current number of rows.
-    public function getIdNum() {
+    public function getIdNum()
+    {
         $botRowCount = $this->db->count_all('assembledbots');
-        $idNum = $botRowCount+1;
+        $idNum = $botRowCount + 1;
         return $idNum;
     }
     
@@ -50,9 +68,9 @@ class Robotsdata extends CI_Model {
     }
 }
 
+    public function deleteAll()
+    {
+        $this->db->empty_table('assembledbots');
+    }
 
-
-
-
-
-
+}