+++ changelog.md
@@ -81,3 +81,10 @@
 
 ### *v.023* (April 1, 2017)
 - Add API key function (Jake)
+
+### *v.024* (April 1, 2017)
+- Added the reboot function (Karan)
+
+### *v.025* (April 1, 2017)
+- Created history page (Jake)
+- Added filters, navigation, order functions on history page (Jake)